+++ src/main/kotlin/com/github/bgomar/consolelogger/UpdateNotify.kt
@@ -1,11 +1,11 @@
 package com.github.bgomar.consolelogger
 
+import com.github.bgomar.bgconsolelogger.tools.ConsoleLoggerSettings
 import com.intellij.ide.plugins.PluginManagerCore
 import com.intellij.notification.NotificationType
 import com.intellij.openapi.extensions.PluginId
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.startup.StartupActivity
-import com.github.bgomar.bgconsolelogger.tools.ConsoleLoggerSettings
 
 class UpdateNotify : StartupActivity {
   private val plugin = PluginManagerCore.getPlugin(PluginId.getId("com.github.bgomar.consolelogger"))!!
@@ -14,6 +14,7 @@ class UpdateNotify : StartupActivity {
     val settings = ConsoleLoggerSettings.getInstance();
     if (settings.version == "Unknown") {
       settings.version = plugin.version
+      showUpdate(project)
     } else if (plugin.version != settings.version) {
       settings.version = plugin.version
       showUpdate(project)
@@ -23,8 +24,11 @@ class UpdateNotify : StartupActivity {
   private val updateContent: String by lazy {
     //language=HTML
     """
-     v0.0.24 Code Refactored, side window design upgraded.<br/>
-
+    🐛 Bugfix for Saving 😁👌<br><br>
+    Sorry it took so long, it was quite a hard task.<br>
+    🤏 Small changes to defaults: <br> 
+    - changed some of the input field sizes
+    - added option for second preset of loggers
     """
   }